requirements.txt
kafka-python
aiohttp
aiofiles
